pyproject.toml
[tool.poetry]
name = "docker-test"
version = "0.1.0"
description = ""
authors = ["Vytautas Pliadis <vytautas.pliadis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.0"
pydantic = "^2.6.1"
sqlalchemy = "^2.0.27"
python-dotenv = "^1.0.1"
poetry-plugin-export = "^1.6.0"
psycopg2-binary = "^2.9.9"
numpy = "^1.26.4"
kaggle = "^1.6.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
